perf: isolate composer draft state in memoized Composer
Move the chat composer (draft textarea, toolbar, model/reasoning
menus, slash/mention popovers, attachments, voice UI) out of ChatView
into a memoized Composer component that owns the draft text as local
state, so every keystroke no longer re-renders App/ChatView.

- Composer owns draft, attachments, voice/STT, slash + mention
  filtering (via profiles prop), model/reasoning menus; submits via
  explicit-text submit(attachments, text) / submitVoice(text)
- ChatView keeps thread/header/error banner/drag shell and renders
  Composer with stable callbacks; edit-message and shell drop routing
  via editRequest state and a drop-files ref; model label mirrored up
- App drops draft/mentions state; submit/stop/openSession/refresh are
  useCallback-stabilized over refs so Composer memo holds during
  streaming
